fix(tests): lsp friendly test cleanup
Closes RNDSTROPPY-56
